Routine Development. -Bug fixes for ScrudFactory project. -Created MenuAccess Policy -Others

+++ FrontEnd/MixERP.Net.FrontEnd/db/src/02.functions-and-logic/core/core.get_shipping_address_by_shipping_address_id.sql
@@ -1,6 +1,6 @@
-DROP FUNCTION IF EXISTS core.get_shipping_address_by_shipping_address_id(integer);
+DROP FUNCTION IF EXISTS core.get_shipping_address_by_shipping_address_id(bigint);
 
-CREATE FUNCTION core.get_shipping_address_by_shipping_address_id(integer)
+CREATE FUNCTION core.get_shipping_address_by_shipping_address_id(bigint)
 RETURNS text
 AS
 $$
